[AARCH64] fix sign-conversion

+++ linux_hello_world.c
@@ -107,7 +107,7 @@ static long sys_write(unsigned int fd, const char *buf, unsigned long len)
     register long x8 __asm__("x8") = 64; /* syscall number: write */
     register long x0 __asm__("x0") = fd;
     register const char *x1 __asm__("x1") = buf;
-    register long x2 __asm__("x2") = len;
+    register long x2 __asm__("x2") = (long)len;
     __asm__ volatile(
         "svc 0"
         : "+r"(x0)
